+++ sigma/modules/moderation/server_settings/starboard/starboardchannel.py
@@ -27,17 +27,14 @@ async def starboardchannel(cmd, pld):
     :type pld: sigma.core.mechanics.payload.CommandPayload
     """
     if pld.msg.author.permissions_in(pld.msg.channel).manage_guild:
-        if pld.msg.channel_mentions:
-            target_channel = pld.msg.channel_mentions[0]
-            if pld.msg.guild.me.permissions_in(target_channel).send_messages:
-                starboard_doc = pld.settings.get('starboard') or {}
-                starboard_doc.update({'channel_id': target_channel.id})
-                await cmd.db.set_guild_settings(pld.msg.guild.id, 'starboard', starboard_doc)
-                response = ok(f'Starboard channel set to {target_channel.name}.')
-            else:
-                response = error('I can\'t write in that channel.')
+        target = pld.msg.channel_mentions[0] if pld.msg.channel_mentions else pld.msg.channel
+        if pld.msg.guild.me.permissions_in(target).send_messages:
+            starboard_doc = pld.settings.get('starboard') or {}
+            starboard_doc.update({'channel_id': target.id})
+            await cmd.db.set_guild_settings(pld.msg.guild.id, 'starboard', starboard_doc)
+            response = ok(f'Starboard channel set to {target.name}.')
         else:
-            response = error('No channel targeted.')
+            response = error('I can\'t write in that channel.')
     else:
         response = denied('Access Denied. Manage Server needed.')
     await pld.msg.channel.send(embed=response)